backend/requirements-baremetal.txt
# Bare-metal installation requirements (non-Docker)
# Includes PyTorch CPU and sentence-transformers

# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL adapter (optional, for PostgreSQL support)
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0

# Testing (optional for development)
pytest==7.4.3
pytest-asyncio==0.21.1

# LLM and AI dependencies
litellm==1.40.0

# PyTorch CPU-only (must be installed separately before this file)
# Install with: pip install --index-url https://download.pytorch.org/whl/cpu torch==2.6.0+cpu torchaudio==2.6.0
# Do NOT install from PyPI as it includes GPU dependencies (~900MB vs ~150MB)
# torch and torchaudio are installed separately in install.sh

# Semantic memory and embeddings
pgvector==0.3.6
sentence-transformers==2.3.0
tiktoken==0.5.2
numpy==1.26.4
huggingface-hub>=0.34.0,<1.0  # Required for sentence-transformers (0.20.0 was too old)

# TTS Dependencies
websockets==12.0  # For VibeVoice WebSocket support

# STT Dependencies
faster-whisper==1.0.3  # Optimized Whisper inference
webrtcvad==2.0.10  # Voice activity detection (legacy)
# torchaudio installed separately with PyTorch CPU (see install.sh)
# silero-vad installed via direct download in service (no pip package)

